Rename name parameter to path
It is not necessary for a type to have an explicitly defined name
parameter. In order to align with other file-based types, it is more
intuitive to use a parameter named "path" to specify where in the local
filesystem a new file should be created, and this should be the namevar
of the type.

This commit changes the namevar to path. It is backwards-compatible with
usage of the "name" parameter due to how namevar works.