Modify style to limit line length to 80 characters
1928 IBM punchcards had 80 columns, largely due to the punchcards having
been designed (in the late 1880s) to be the same size as currency so as
to be able to carry them around in currency carriers. It is obvious
therfore, for these reasons, that we should today try and write code
that has no more than 80 characters per line.

This commit modifies the provider ruby code to conform to the horizontal
width of a late 1880s currency carrier, were it to be printed monospace
in an appropriate font size.