Generalize http method logic, use tempfile
In preparation for performing HTTP HEAD requests, generalize the method
containing the Net::HTTP logic. This commit moves the logic from the
http_get() method to a new method, http(), which accepts an IO object to
write the body to and returns the HTTP response object for any request
made. It supports performing HEAD or other request types via the options
parameter.

While making these changes it was convenient to introduce a tempfile
download pattern, wherein rather than saving directly to the target
path, the provider will create a tempfile and mv it to the target path
upon successful download. This pattern prevents incomplete files from
being dropped into the target path.