Add support for a proxy parameter that takes a URI
This commit introduces an alternate means of specifying a proxy,
allowing a URI to be given instead of separately specified host, port,
and optional username/password.

The value of supporting this kind of alternative specification is if
organizations are already storing such information in e.g. Hiera as a
single URI string, they can pass it directly to the type instead of
parsing it out with regexp or similar to separate the host from the
port.