Add LICENSE file, version support to metadata
In order for the module to show up in all appropriate filtered searches
on the Forge the metadata needs to contain OS supported version
information, as well as Puppet / Puppet Enterprise supported version
information. This commit adds that additional metadata.

It also adds a copy of the Apache 2.0 license to the module root.