Convert module to PDK
The PDK provides a standard framework for style, building and running
tests, and maintaining metadata. While there are no tests currently
written for this module, converting to the PDK lowers the barrier to
entry for writing them should anyone want to do that in the future.

As part of converting to the PDK many Ruby style edits were made in
order to align with the conventional rubocop linting and style tests.

Because Ruby 1.9.x conventions are used, the supported version of Puppet
is changed from >=3.x to >=4.6.x. Older versions may work but it is
difficult to find documentation for versions of Puppet older than 4.6.
The oldest version of Ruby supported by Puppet 4.6 is 1.9.3.